# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Checker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Resume {
  id: number;
  userId: number;
  title: string;
  content: string;
  /** @nullable */
  overallScore: number | null;
  /** @nullable */
  atsScore: number | null;
  /** @nullable */
  jobRole: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateResumeBody {
  title: string;
  content: string;
  /** @nullable */
  jobRole?: string | null;
}

export interface AnalyzeResumeBody {
  jobRole?: string;
  jobDescription?: string;
}

export interface Skill {
  name: string;
  level: string;
  category: string;
}

export interface MissingSkill {
  name: string;
  importance: string;
  reason: string;
}

export interface CareerPath {
  role: string;
  match: number;
  description: string;
}

export interface ResumeAnalysis {
  resumeId: number;
  overallScore: number;
  atsScore: number;
  toneScore: number;
  clarityScore: number;
  impactScore: number;
  detectedSkills: Skill[];
  missingSkills: MissingSkill[];
  careerPaths: CareerPath[];
  strengths: string[];
  weaknesses: string[];
  summary: string;
  /** @nullable */
  jobRoleMatch: number | null;
  toneAnalysis: string;
}

export interface EnhancementItem {
  section: string;
  issue: string;
  suggestion: string;
  priority: string;
  /** @nullable */
  originalText: string | null;
  /** @nullable */
  improvedText: string | null;
}

export interface ResumeEnhancement {
  resumeId: number;
  enhancements: EnhancementItem[];
  /** @nullable */
  rewrittenSummary: string | null;
  keywordsToAdd: string[];
  formattingTips: string[];
}

export interface InterviewQuestion {
  question: string;
  category: string;
  difficulty: string;
  tip: string;
}

export interface InterviewQuestions {
  resumeId: number;
  questions: InterviewQuestion[];
}

export interface DashboardStats {
  totalResumes: number;
  /** @nullable */
  averageScore: number | null;
  /** @nullable */
  bestScore: number | null;
  recentActivity: Resume[];
}