# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Checker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register new user
 */
export const RegisterBody = zod.object({
  email: zod.string(),
  password: zod.string(),
  name: zod.string(),
});

/**
 * @summary Login user
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
    createdAt: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout user
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary List user's resumes
 */
export const GetResumesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string(),
  content: zod.string(),
  overallScore: zod.number().nullable(),
  atsScore: zod.number().nullable(),
  jobRole: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetResumesResponse = zod.array(GetResumesResponseItem);

/**
 * @summary Upload and create resume
 */
export const CreateResumeBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  jobRole: zod.string().nullish(),
});

/**
 * @summary Get resume by ID
 */
export const GetResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResumeResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string(),
  content: zod.string(),
  overallScore: zod.number().nullable(),
  atsScore: zod.number().nullable(),
  jobRole: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete resume
 */
export const DeleteResumeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Analyze resume with AI
 */
export const AnalyzeResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const AnalyzeResumeBody = zod.object({
  jobRole: zod.string().optional(),
  jobDescription: zod.string().optional(),
});

export const AnalyzeResumeResponse = zod.object({
  resumeId: zod.number(),
  overallScore: zod.number(),
  atsScore: zod.number(),
  toneScore: zod.number(),
  clarityScore: zod.number(),
  impactScore: zod.number(),
  detectedSkills: zod.array(
    zod.object({
      name: zod.string(),
      level: zod.string(),
      category: zod.string(),
    }),
  ),
  missingSkills: zod.array(
    zod.object({
      name: zod.string(),
      importance: zod.string(),
      reason: zod.string(),
    }),
  ),
  careerPaths: zod.array(
    zod.object({
      role: zod.string(),
      match: zod.number(),
      description: zod.string(),
    }),
  ),
  strengths: zod.array(zod.string()),
  weaknesses: zod.array(zod.string()),
  summary: zod.string(),
  jobRoleMatch: zod.number().nullable(),
  toneAnalysis: zod.string(),
});

/**
 * @summary Get enhancement suggestions
 */
export const EnhanceResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const EnhanceResumeResponse = zod.object({
  resumeId: zod.number(),
  enhancements: zod.array(
    zod.object({
      section: zod.string(),
      issue: zod.string(),
      suggestion: zod.string(),
      priority: zod.string(),
      originalText: zod.string().nullable(),
      improvedText: zod.string().nullable(),
    }),
  ),
  rewrittenSummary: zod.string().nullable(),
  keywordsToAdd: zod.array(zod.string()),
  formattingTips: zod.array(zod.string()),
});

/**
 * @summary Generate interview questions based on resume
 */
export const GetInterviewQuestionsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetInterviewQuestionsResponse = zod.object({
  resumeId: zod.number(),
  questions: zod.array(
    zod.object({
      question: zod.string(),
      category: zod.string(),
      difficulty: zod.string(),
      tip: zod.string(),
    }),
  ),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalResumes: zod.number(),
  averageScore: zod.number().nullable(),
  bestScore: zod.number().nullable(),
  recentActivity: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      title: zod.string(),
      content: zod.string(),
      overallScore: zod.number().nullable(),
      atsScore: zod.number().nullable(),
      jobRole: zod.string().nullable(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
});