# lib/api-zod/src/generated/types/careerPath.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Resume Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerPath {
  role: string;
  match: number;
  description: string;
}